Countinued work on prototype
New additions to the prototype include:

. A spring joint function that allows the player to draw back the bird
from an anchored point on the rotating catapult.

. A delayed release time that allows the bird to be released from the
anchor point upon the player releasing the mouse key, allowing the bird
to accelerate upwards at the draw angle instead of snapping back to the
anchor point.

.Updated placeholder graphics that allow team members differentiate
between gameobjects.

.A secondary camera that flows the distance traveled by the bird. The
main camera is deactivated whilst the secondary camera is activated once
the bird is launched.

.A stamina bar standing at 250 x 20 has been added to the upper right of
the UI canvas, although this stamina bar is not operational yet.