Created new build an completed implementing and coding wind functionality.
• The game stores seven different wind variables, these variables range
from: -3, -2, -1, 0, 1, 2, 3.
• -3, -2 and - 1 attain a negative wind direction meaning the wind
created from these variables will push the player to the left whilst the
wind created from 1, 2 and 3 attain positive wind directions and will
push the birds further to the right.
• -3 and 3 have the highest of their respected power variables whilst -1
and 1 have the lowest.
• 0 attains no wind variable and is the default setting for our game.

• When the game is launched one of a wind variable is randomly
generated.
• After 2.5 seconds the force and direction of the wind will begin to
gradually decrease or increase depending on the variable generated at
random.

Whilst implementing the functionality I noticed a slight issue. As the
power and direction of the wind is always fluctuating the player’s bird
in turn never comes to a standstill (the round never ends). To combat
this issue I had implemented a functionality in our code that would
revert the wind power and direction to its default value of zero after
45 seconds of gameplay, allowing us to avoid a never ending round.

Currently the only visual feedback that suggest to the fluctuation of
wind are the “Wind direction” arrows found in our UI. Whilst working on
the task I experimented with implemented particles as visual feedback
however, during the implementations of the particles I found that the
birds were traveling faster than the particles could render on screen.